+++ src/locales/en.ts
@@ -8,8 +8,16 @@ export const en = {
   Theme: "Theme",
   Light: "Light",
   Dark: "Dark",
+  SystemColor: "System color",
   PageNotFound: "Page not found",
   BackToLandingPage: "Back to home",
+  BuildInfo: "Build info",
+  Contributors: "Contributors",
+  Version: "Version",
+  PageTitle: "My App",
+  PageTitleLandingPage: "Home",
+  PageTitleChangelog: "Changelogs",
+  PageTitleError404: "Page not found",
 };
 
 export type Localization = typeof en;