+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+`2025-03-01` - `0.1.0`
+- `Feature`: Support for Async Components
+- `Feature`: System Color Mode
+- `Change`: Improved Logo Button
+- `Misc`: Dependencies updated to latest versions
+
 `2025-02-02` - `0.0.3`
 - `Misc`: Dependencies updated to latest versions